+++ src/models/VecTvpGamma.cpp
@@ -9,6 +9,7 @@
 // reading the input, and putting the results back.
 
 #include "models/models.h"
+#include "models/model_check.h"
 
 #include "bayests/vec_tvp_gamma.h"
 #include "io/hdf5/hdf5_and_armadillo.h"
@@ -123,4 +124,9 @@ void VecTvpGamma::log_likelihood(const ModelLocation &location_arg)
         bayests::VecTvpGammaSampler{}.log_likelihood(input, draws);
 
     bayests::hdf5_io::write_log_likelihood(file, loglik);
+}
+
+ModelCheck VecTvpGamma::check(const ModelLocation &location_arg)
+{
+    return check_vec_model(location_arg, [](const ModelFile &file) { return io::read_input(file); });
 }